05_Firestore/webhook/requirements.txt
line-bot-sdk
google-cloud-dialogflow
protobuf-to-dict
firebase_admin
google-cloud-pubsub
google-cloud-secret-manager==2.1.0